+++ spec/features/users_can_view_a_project_spec.rb
@@ -15,7 +15,7 @@ RSpec.feature "Users can view a project" do
     within("#project-summary") do
       expect(page).to have_content(project.incoming_trust.name)
       expect(page).to have_content(project.provisional_conversion_date.to_formatted_s(:govuk))
-      expect(page).to have_content(project.establishment.local_authority)
+      expect(page).to have_content(project.establishment.local_authority_name)
       expect(page).to have_content(project.incoming_trust.name)
       expect(page).to have_content(project.establishment.region_name)
       expect(page).to have_link(I18n.t("project.summary.establishment_sharepoint_link.value"), href: project.establishment_sharepoint_link)