+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - new MembersApi Client to call the Parliamentary Members API and retrieve MP
   details
 - show the Member of Parliament details for a school on a new page
+- Users can view a project's local authority details from the project
+  information tab
 
 ### Changed